Rename, Remove: 컴포넌트 및 이미지 정리
