+++ src/components/common/TripTitle.tsx
