+++ src/pages/tripplace/index.tsx
@@ -2,9 +2,9 @@ import TripPlaceLayout from "@/layout/tripplace/layout";
 import HeaderTripSelect from "@/components/header/HeaderTripSelect";
 import TripPlacesMap from "@/components/tripplace/TripPlacesMap";
 import TripRegionDaysEdit from "@/components/common/TripRegionDaysEdit";
-import SelectDateInfo from "@/components/tripselect/SelectDateInfo";
-import SelectItem from "@/components/tripselect/SelectItem";
-import ButtonLarge from "@/components/tripselect/ButtonLarge";
+import SelectDateInfo from "@/components/common/SelectDateInfo";
+import SelectItem from "@/components/common/SelectItem";
+import ButtonLarge from "@/components/common/ButtonLarge";
 import { PlacesStore } from "@/store/PlacesStore";
 import DefaultImage from "@/components/common/DefaultImage";
 import LocalPlaceItem from "@/components/tripplace/LocalPlaceItem";
